feat: extract top bar actions to a component
